+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.0.2
+
+* **⚠️ Breaking:** Added lazy collection path support, collection path is now a callback so you may use dynamic id's in the path if needed and the API will stay in sync.
+* **✨ New:** Added support for collection group queries with `FirestoreAPI._isCollectionGroup`.
+
+## 0.0.1+5
+
+* Fixed bug where feedback response required for Writeable.isValid method required generic specifications.
+
 ## 0.0.1+4
 
 * Update example project and other small changes.